+++ Code/Mantid/Framework/Algorithms/src/FindDetectorsOutsideLimits.cpp
@@ -94,7 +94,7 @@ namespace Mantid
       // Keep track of those reading low/high
       size_t numLow(0), numHigh(0);
 
-      for (int i = 0; i < numSpec; ++i)
+      for (int64_t i = 0; i < int64_t(numSpec); ++i)
       {
         if (i % progStep == 0)
         {
